backend/requirements.txt
fastapi
uvicorn
chromadb
langchain
langchain-community
langchain-google-genai
langchain-ollama
watchdog
python-multipart
websockets
pydantic
python-dotenv
